+++ src/apps/settings/modules/seelenweg/infra.tsx
@@ -170,6 +170,20 @@ export const SeelenWegSettings = () => {
             />
           </SettingsOption>
           <SettingsOption>
+            <div>{t('weg.items.thumbnail_generation_enabled')}</div>
+            <Switch
+              checked={settings.thumbnailGenerationEnabled}
+              onChange={(value) => dispatch(SeelenWegActions.setThumbnailGenerationEnabled(value))}
+            />
+          </SettingsOption>
+          <SettingsOption>
+            <div>{t('weg.items.show_instance_counter')}</div>
+            <Switch
+              checked={settings.showInstanceCounter}
+              onChange={(value) => dispatch(SeelenWegActions.setShowInstanceCounter(value))}
+            />
+          </SettingsOption>
+          <SettingsOption>
             <div>{t('weg.items.visible_separators')}</div>
             <Switch
               checked={settings.visibleSeparators}