+++ src/background/modules/cli/mod.rs
@@ -12,7 +12,7 @@ use application::{get_app_command, handle_cli_events};
 pub use domain::SvcAction;
 use domain::SvcMessage;
 use itertools::Itertools;
-use tauri_plugin_dialog::{DialogExt, MessageDialogKind};
+use tauri_plugin_dialog::{DialogExt, MessageDialogButtons, MessageDialogKind};
 use windows::Win32::{
     System::TaskScheduler::{IExecAction2, ITaskService, TaskScheduler},
     UI::Shell::FOLDERID_LocalAppData,
@@ -157,7 +157,12 @@ impl ServiceClient {
     fn start_service_task() -> Result<()> {
         Com::run_with_context(|| unsafe {
             let task_service: ITaskService = Com::create_instance(&TaskScheduler)?;
-            task_service.Connect(None, None, None, None)?;
+            task_service.Connect(
+                &Default::default(),
+                &Default::default(),
+                &Default::default(),
+                &Default::default(),
+            )?;
             let folder = task_service.GetFolder(&"\\Seelen".into())?;
             let task = folder.GetTask(&"Seelen UI Service".into())?;
 
@@ -170,7 +175,7 @@ impl ServiceClient {
             let service_path = Self::service_path()?.to_string_lossy().to_lowercase();
             match action_path.to_string().to_lowercase() == service_path {
                 true => {
-                    task.Run(None)?;
+                    task.Run(&Default::default())?;
                     Ok(())
                 }
                 false => {
@@ -198,7 +203,9 @@ impl ServiceClient {
                     .message(t!("service.not_running_description"))
                     .title(t!("service.not_running"))
                     .kind(MessageDialogKind::Info)
-                    .ok_button_label(t!("service.not_running_ok"))
+                    .buttons(MessageDialogButtons::OkCustom(
+                        t!("service.not_running_ok").to_string(),
+                    ))
                     .blocking_show();
                 if try_again {
                     script.execute().await?;