+++ src/background/seelen_bar/mod.rs
@@ -11,11 +11,14 @@ use crate::{
         are_overlaped,
         constants::{NATIVE_UI_POPUP_CLASSES, OVERLAP_BLACK_LIST_BY_EXE},
     },
-    windows_api::{window::Window, AppBarData, AppBarDataEdge, WindowsApi},
+    windows_api::{window::Window, AppBarData, WindowsApi},
 };
 use base64::Engine;
 use itertools::Itertools;
-use seelen_core::{handlers::SeelenEvent, state::HideMode};
+use seelen_core::{
+    handlers::SeelenEvent,
+    state::{FancyToolbarSide, HideMode},
+};
 use serde::Serialize;
 use tauri::{Emitter, Listener, WebviewWindow};
 use windows::Win32::{
@@ -36,7 +39,7 @@ pub struct FancyToolbar {
 impl Drop for FancyToolbar {
     fn drop(&mut self) {
         log::info!("Dropping {}", self.window.label());
-        if let Ok(hwnd) = self.window.hwnd() {
+        if let Ok(hwnd) = self.hwnd() {
             AppBarData::from_handle(hwnd).unregister_bar();
         }
         log_error!(self.window.destroy());
@@ -44,6 +47,10 @@ impl Drop for FancyToolbar {
 }
 
 impl FancyToolbar {
+    pub fn hwnd(&self) -> Result<HWND> {
+        Ok(HWND(self.window.hwnd()?.0))
+    }
+
     pub fn new(monitor: &str) -> Result<Self> {
         Ok(Self {
             window: Self::create_window(monitor)?,
@@ -92,7 +99,7 @@ impl FancyToolbar {
             } else if let Some(past_window) = self.last_overlapped_window {
                 if past_window != window
                     && past_window.monitor() != monitor
-                    && Window::from(self.window.hwnd()?).monitor() != monitor
+                    && Window::from(self.hwnd()?).monitor() != monitor
                 {
                     return Ok(());
                 }
@@ -105,7 +112,7 @@ impl FancyToolbar {
     }
 
     pub fn hide(&mut self) -> Result<()> {
-        WindowsApi::show_window_async(self.window.hwnd()?, SW_HIDE)?;
+        WindowsApi::show_window_async(self.hwnd()?, SW_HIDE)?;
         self.window.emit_to(
             self.window.label(),
             SeelenEvent::HandleLayeredHitboxes,
@@ -115,7 +122,7 @@ impl FancyToolbar {
     }
 
     pub fn show(&mut self) -> Result<()> {
-        WindowsApi::show_window_async(self.window.hwnd()?, SW_SHOWNOACTIVATE)?;
+        WindowsApi::show_window_async(self.hwnd()?, SW_SHOWNOACTIVATE)?;
         self.window.emit_to(
             self.window.label(),
             SeelenEvent::HandleLayeredHitboxes,
@@ -147,21 +154,32 @@ impl FancyToolbar {
     /// so we use this functions that only takes the toolbar in account
     pub fn get_work_area_by_monitor(monitor: HMONITOR) -> Result<RECT> {
         let monitor_info = WindowsApi::monitor_info(monitor)?;
-
         let dpi = WindowsApi::get_device_pixel_ratio(monitor)?;
         let mut rect = monitor_info.monitorInfo.rcMonitor;
-
         let state = FULL_STATE.load();
+        let settings = state.settings().fancy_toolbar();
         if state.is_bar_enabled() {
-            let toolbar_height = state.settings().fancy_toolbar().height;
-            rect.top += (toolbar_height as f32 * dpi) as i32;
+            match settings.position {
+                FancyToolbarSide::Top => {
+                    rect = RECT {
+                        top: rect.top + (settings.height as f32 * dpi) as i32,
+                        ..rect
+                    };
+                }
+                FancyToolbarSide::Bottom => {
+                    rect = RECT {
+                        bottom: rect.bottom - (settings.height as f32 * dpi) as i32,
+                        ..rect
+                    };
+                }
+            }
         }
 
         Ok(rect)
     }
 
     pub fn set_position(&mut self, monitor: HMONITOR) -> Result<()> {
-        let hwnd = HWND(self.window.hwnd()?.0);
+        let hwnd = HWND(self.hwnd()?.0);
 
         let state = FULL_STATE.load();
         let settings = &state.settings().fancy_toolbar();
@@ -169,15 +187,26 @@ impl FancyToolbar {
         let monitor_info = WindowsApi::monitor_info(monitor)?;
         let monitor_dpi = WindowsApi::get_device_pixel_ratio(monitor)?;
         let rc_monitor = monitor_info.monitorInfo.rcMonitor;
-        self.theoretical_rect = RECT {
-            bottom: rc_monitor.top + (settings.height as f32 * monitor_dpi) as i32,
-            ..rc_monitor
-        };
+
+        match settings.position {
+            FancyToolbarSide::Top => {
+                self.theoretical_rect = RECT {
+                    bottom: rc_monitor.top + (settings.height as f32 * monitor_dpi) as i32,
+                    ..rc_monitor
+                };
+            }
+            FancyToolbarSide::Bottom => {
+                self.theoretical_rect = RECT {
+                    top: rc_monitor.bottom - (settings.height as f32 * monitor_dpi) as i32,
+                    ..rc_monitor
+                };
+            }
+        }
 
         let mut abd = AppBarData::from_handle(hwnd);
         match settings.hide_mode {
             HideMode::Never => {
-                abd.set_edge(AppBarDataEdge::Top);
+                abd.set_edge(settings.position.into());
                 abd.set_rect(self.theoretical_rect);
                 abd.register_as_new_bar();
             }